requirements.txt
numpy<2
pillow==10.4.0
torch==2.3.0
torchvision==0.18.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
